move methods to base class
